// vivah-ui/src/app/profile/models/partner-preference.ts
export interface PartnerPreference {
    IdGender: number;
    MinAge: number;
    MaxAge: number;
    IdMaritalStatus: number;
    MinHieght: number;
    MaxHeight: number;
    IdEatingHabit: number;
    IdDrinkingHabit: number;
    IdSmokingHabit: number;
    IdMotherTongue: number;
    IdReligion: number;
    IdCaste: number;
    IdSubCaste: number;
    IdGothra: number;
    IdNakshatra: number;
    IdRashi: number;
    IdQualification: number;
    IdEmployement: number;
    IdPersonalDetails: number;
}